# src/main/api-spec/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    CarBE:
      type: object
      properties:
        id:
          type: integer
          format: int64
        model:
          type: string
        brand:
          type: string
        date:
          $ref: "#/components/schemas/Instant"
      required:
      - model
      - brand
    CarCheckMappingRequest:
      type: object
      required:
      - model
      - brand
      properties:
        model:
          type: string
        brand:
          type: string
        date:
          $ref: "#/components/schemas/Instant"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
paths:
  /cars:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarCheckMappingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarBE"
      summary: Replace Car
      tags:
      - Car Resource
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarBE"
      summary: All
      tags:
      - Car Resource
  /cars/brand/{brand}:
    delete:
      parameters:
      - name: brand
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      summary: Delete Car By Brand
      tags:
      - Car Resource
  /cars/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarBE"
      summary: Get
      tags:
      - Car Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      summary: Delete Car
      tags:
      - Car Resource
info:
  title: gitpod-example API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value